refactor(frontend): add useCopyToClipboard hook
